==> src/01_scrape_readmes/02_scrape_readmes.ipynb <==
### GitHub README Extraction 

During this summer, the DSGP OSS 2021 team willbe classifying GitHub repositores into different software types. To do this, we will be extracting README files from all of the OSS repos (i.e. those with OSS licenses) and then developing NLP techniques to classify those repos. In this file, we document the extraction process for GitHub README files. 

First, we load our packages 

# load packages 
import os
import glob
import psycopg2 as pg
from sqlalchemy import create_engine
import pandas as pd
import requests as r
import string 
import json
import base64
import urllib.request
import itertools 
from bs4 import BeautifulSoup
from dotenv import load_dotenv
import warnings
warnings.simplefilter(action='ignore', category=FutureWarning)

import time
from functools import wraps
import urllib.request
import multiprocessing
from multiprocessing.pool import ThreadPool as Pool
from retrying import retry
from datetime import datetime

print("ready")

Next, we will grab our data from the database. 

# connect to the database, download data 
connection = pg.connect(host = 'postgis1', database = 'sdad', 
                        user = os.environ.get('db_user'), 
                        password = os.environ.get('db_pwd'))

raw_slug_data = '''SELECT * FROM gh_2007_2020.repos_ranked where commits > 1000  '''

# convert to a dataframe, show how many missing we have (none)
raw_slug_data = pd.read_sql_query(raw_slug_data, con=connection)
raw_slug_data.head()

raw_slug_data['slug'].count

batch = "batch3"

raw_slugs = raw_slug_data["slug"].tolist()
slugs = []
for s in raw_slugs:
    slugs.append(s.strip())

# exp multiplier that waits 30 seconds, then 30 secs + 30 secs, etc until 5 tries is over 
@retry(stop_max_attempt_number=5, wait_exponential_multiplier=30000, wait_exponential_max=30000) 
def scrape_slugs(slug):
    ''' Scrapes slugs with multiple threads '''
    url = f'https://github.com/{slug}#readme'
    split_slugs = slug.split("/")
    login = split_slugs[0]
    repo = split_slugs[1]
    fullfilename = os.path.join('/project/class/bii_sdad_dspg/ncses_oss_2021/', f'readme_{login}_{repo}.txt')
    urllib.request.urlretrieve(url, fullfilename)
    now = datetime.now()
    current_time = now.strftime("%Y-%m-%d %H:%M:%S")
    return slug, current_time, "Done"

cores_available = multiprocessing.cpu_count() - 1
pool = Pool(cores_available)

logging_slugs = []
logging_time = []
logging_status = []
for result in pool.imap_unordered(scrape_slugs, slugs):
    logging_slugs.append(result[0])
    logging_time.append(result[1])
    logging_status.append(result[2])
    logging_df = pd.DataFrame({'slug': logging_slugs, "batch": batch, 'as_of': logging_time, 'status': logging_status}, 
                              columns=["slug", "batch", "as_of", "status"])
    logging_df.to_csv('/project/class/bii_sdad_dspg/ncses_oss_2021/'+batch+'.csv', index=False)
    print(result)

pd.read_csv('data.csv')



slugs = ["brandonleekramer/diversity", 
         "uva-bi-sdad/oss-2020", 
         "facebook/react", 
         "RichardLitt/standard-readme"
] 



slugs = ["brandonleekramer/diversity", "uva-bi-sdad/oss-2020", "facebook/react", "RichardLitt/standard-readme"] #test data 
#slugs = raw_slug_data.slug.tolist()

#myPath = '/sfs/qumulo/qhome/kb7hp/git/oss-2020/src/09_repository-scraping/' 
myPath = '/project/class/bii_sdad_dspg/ncses_oss_2021/'

for slug in slugs:
    url = f'https://github.com/{slug}#readme'
    split_slugs = slug.split("/")
    login = split_slugs[0]
    repo = split_slugs[1]
    fullfilename = os.path.join(myPath, f'readme_{login}_{repo}.txt')
    urllib.request.urlretrieve(url, fullfilename)
    print(f'Finished scraping: {login}/{repo}')

Note to Crystal: 

The function above this note works for small-scale scraping but we need to add in the rate limit on API calls before we scale up. 
https://stackoverflow.com/questions/40748687/python-api-rate-limiting-how-to-limit-api-calls-globally

We could also try to add in multiprocessing to speed things up. I'm not sure this link is the right one, but we can chat more about that.
https://stackoverflow.com/questions/54858979/how-to-use-multiprocessing-with-requests-module

myPath = '/project/class/bii_sdad_dspg/ncses_oss_2021/'

repo_name = []
readme_text = [] 
for filename in os.listdir(myPath):
    if filename.endswith('.txt'):
        with open(os.path.join(myPath, filename)) as f:
            content = f.read()
            soup = BeautifulSoup(content, 'html.parser')
            if soup.article is not None:
                clean_html = ''.join(soup.article.findAll(text=True))
            else: 
                clean_html = None 
            repo_name.append(filename)
            readme_text.append(clean_html)
            df = pd.DataFrame({'slug': repo_name, 'readme_text': readme_text}, columns=["slug", "readme_text"])
            df['slug'] = df['slug'].str.replace('readme_','')
            df['slug'] = df['slug'].str.replace('.txt','')
            # this works because slugs can't have underscores
            df['slug'] = df['slug'].str.replace('_','/') 
df 

os.chdir('/project/class/bii_sdad_dspg/ncses_oss_2021/')
extension = 'csv'
all_filenames = [i for i in glob.glob('*.{}'.format(extension))]
combined_csv = pd.concat([pd.read_csv(f) for f in all_filenames ])
merged_df = df.merge(combined_csv, how='inner', on='slug')
merged_df

os.chdir('/sfs/qumulo/qhome/kb7hp/git/dspg21oss/data/')
merged_df.to_csv("readme_test_data.csv")

db_user = os.environ.get('db_user')
db_pwd = os.environ.get('db_pwd')
engine = create_engine('postgresql+psycopg2://{db_user}:{db_pwd}@postgis1/sdad')
merged_df.to_sql('repos_scraped', engine, schema='gh_2007_2020', if_exists='replace', index=False)
print("It worked")

We need to write this to the database now... 

Try this: https://medium.com/analytics-vidhya/part-3-5-pandas-dataframe-to-postgresql-using-python-d3bc41fcf39 

https://pybay.com/site_media/slides/raymond2017-keynote/process.html

https://github.com/rholder/retrying 

https://stackoverflow.com/questions/45016816/updating-row-using-sqlalchemy-and-scrapy

https://stackoverflow.com/questions/2712524/handling-urllib2s-timeout-python

# connect to the database, download data 
connection = pg.connect(host = 'postgis1', 
                        database = 'sdad', 
                        user = os.environ.get('db_user'), 
                        password = os.environ.get('db_pwd'))

repos_scraped = '''SELECT * FROM gh_2007_2020.repos_scraped '''

# convert to a dataframe, show how many missing we have (none)
repos_scraped = pd.read_sql_query(repos_scraped, con=connection)
repos_scraped.head()
