==> src/01_scrape_readmes/03.a2_scrape_repo_stats.ipynb <==
# Scrap repo stats using PyGitHub

### Author: Crystal Zang

This notebook utilized GitHub access tokens (PAT) to scrape GitHub repository statistics such as stargazers, watchers, forks, and topics. One PAT would scrape at a rate of 5000 repositories per hour. Utilizing 36 PATs we would scrape 10,288,063 repositories in about XXX hours at a rate of 15,514 repositories per hour.

#### Warnings
You should not commit any access topen to GitHub, which would result in access token being revoked.


%reset

# load packages 
import os
import psycopg2 as pg
from sqlalchemy import create_engine
import pandas as pd
import requests as r
import string 
import json
import base64
import urllib.request
import itertools 
import numpy as np
from bs4 import BeautifulSoup
from dotenv import load_dotenv
from github import Github, RateLimitExceededException, BadCredentialsException, BadAttributeException, GithubException, UnknownObjectException, BadUserAgentException
import warnings
import datetime

import multiprocessing
#from multiprocessing.pool import ThreadPool as Pool
from multiprocessing import Pool, freeze_support

import concurrent.futures

warnings.simplefilter(action='ignore', category=FutureWarning)

#os.environ['db_user'] = ''
#os.environ['db_pwd'] = ''

# connect to the database, download data, limit to repos with at least 20,000 commits?
connection = pg.connect(host = 'postgis1', database = 'sdad', 
                        user = os.environ.get('db_user'), 
                        password = os.environ.get('db_pwd'))

raw_slug_data = '''SELECT * FROM gh_2007_2020.repos_ranked WHERE (commits BETWEEN '700' AND '800')'''
#raw_slug_data = '''SELECT * FROM gh_2007_2020.repos_ranked WHERE commits < 1000'''

# convert to a dataframe, show how many missing we have (none)
raw_slug_data = pd.read_sql_query(raw_slug_data, con=connection)

connection.close()


print(raw_slug_data.head())
print(raw_slug_data.shape)
print(raw_slug_data.isna().sum())

raw_slug_data = pd.read_csv('/home/zz3hs/git/dspg21oss/data/dspg21oss/crystal_to_scrape_0706.csv') #import csv
raw_slug_data

#get rid of leading and ending space, save slugs to a list
raw_slugs = raw_slug_data["slug"].tolist()
slugs = []
for s in raw_slugs:
    slugs.append(s.strip())  
print(len(slugs))
print(slugs[0], slugs[len(slugs)-1])

#os.environ['db_user'] = ''
#os.environ['db_pwd'] = ''

# connect to the database, download data, limit to repos with at least 20,000 commits?
connection = pg.connect(host = 'postgis1', database = 'sdad', 
                        user = os.environ.get('db_user'), 
                        password = os.environ.get('db_pwd'))

#PATs access token, saved as a dataframe
github_pats = '''SELECT * FROM gh_2007_2020.pats_update'''
github_pats = pd.read_sql_query(github_pats, con=connection)

#PATs access token, saved as a list
access_tokens = github_pats["token"]

#number of tokens available for use, a numeric value
num_token = '''SELECT COUNT(*) FROM gh_2007_2020.pats_update'''
num_token = pd.read_sql_query(num_token, con=connection)
num_token=num_token.iloc[0]['count']

connection.close()
len(access_tokens)

slugs_try = slugs[0:100]
print(slugs_try[0], slugs_try[99])

# index ranges from 0 to maximum number of PATs available
def get_access_token(github_pat_index):
    if github_pat_index < num_token:
       # print("Extracting access token #", github_pat_index+1,", total", num_token, "tokens are available.")
        return github_pats.token[github_pat_index]
    else:
        print("token exceed limit")

#practice on 7 slugs, one invalid slug 
slugs_example = ["moderndive/ModernDive_book", "DSPG-Young-Scholars-Program/dspg21oss", 
                 "unknownrepo", "moderndive/ModernDive_book", "lsst/ip_diffim", "esrlabs/chipmunk", "paulmillr/chokidar"]

# Not using any multiprocessing

def pull_repo_stats(github_pat_index, slugs):
    df_repo_stats = pd.DataFrame()
    for slug in slugs:
        if github_pat_index >= len(access_tokens):
            github_pat_index -= len(access_tokens)
            print("***Pat access token exceed limit, restart access token loop with #", github_pat_index)
        while github_pat_index < len(access_tokens):
            try:
                access_token = get_access_token(github_pat_index)
                #print("Scrapping --", slug,". Extracting access token #", github_pat_index+1,", total", num_token, "tokens are available.")
                #if false, retry until true, max number of retry is 20 times
                g = Github(access_token, retry = 20, timeout = 15)
                repo = g.get_repo(slug)
                df_repo_stats = df_repo_stats.append({
                    "slug": slug,
                    'stars': repo.stargazers_count,
                    'watchers': repo.subscribers_count,
                    'forks': repo.forks_count,
                    'topics': repo.get_topics()
                }, ignore_index = True)
            except RateLimitExceededException as e:
                print(e.status)
                print('Rate limit exceeded --', slug, ", using access token #", github_pat_index)
                print("Current time:", datetime.datetime.now())
                #time.sleep(300)
                github_pat_index+=1
                print("***Exit current access token, proceed with next aceess token #", github_pat_index, "rescrape --",slug)
                break
            except BadCredentialsException as e:
                print(e.status)
                print('Bad credentials exception --', slug, ", using access token #", github_pat_index)
                print("Current time:", datetime.datetime.now())
                github_pat_index+=1
                print("***Exit current access token, proceed with next aceess token #", github_pat_index, "rescrape --",slug)
                break
            except UnknownObjectException as e:
                print(e.status)
                print('Unknown object exception --', slug)
                break
            except GithubException as e:
                print(e.status)
                print('General exception --', slug)
                break
            except r.exceptions.ConnectionError as e:
                print('Retries limit exceeded --', slug)
                print(str(e))
                time.sleep(10)
                continue
            except r.exceptions.Timeout as e:
                print('Time out exception --', slug)
                print(str(e))
                time.sleep(10)
                continue
            break
    return df_repo_stats



slugs1 = slugs[0:10000] #40 minutes
slugs2 = slugs[10001:20000] #38 minutes
slugs3 = slugs[20001:30000]
slugs4 = slugs[30001:40000]
slugs5 = slugs[40001:50000]
slugs6 = slugs[50001:60000]
slugs7 = slugs[60001:70000]
slugs8 = slugs[70001:80000]
slugs9 = slugs[80001:88778]




start_time = datetime.datetime.now()
print("Start scraping:", start_time)
df_repo_stats = pull_repo_stats(6, slugs9)
end_time =  datetime.datetime.now()
print("Finished scraping", len(df_repo_stats), "of", len(slugs9), "records at", end_time)
print("It took", end_time-start_time, "to run.")

# read in the file and check
df_repo_stats.head()
#print(df_repo_stats)
#print(df_repo_stats.isna().sum())
#print(df_repo_stats.shape)


df_repo_stats.to_csv(r'/home/zz3hs/git/dspg21oss/data/dspg21oss/new_repo_stats_9.csv', index = False)   


# Function to scrape one slug using specified pat index

#global github_pat_index
def pull_repo_stats(slug, github_pat_index):
    #github_pat_index = 1
   # if github_pat_index >= len(access_tokens):
   #     github_pat_index -= len(access_tokens)
   #     print("***Pat access token exceed limit, restart access token loop with #", github_pat_index)
    while True:
        try:
            if github_pat_index >= len(access_tokens):
                github_pat_index -= len(access_tokens)
                print(f"***Pat access token exceed limit, restart access token loop with #", github_pat_index)  
                
            access_token = get_access_token(github_pat_index)
            #print("Scrapping --", slug,". Extracting access token #", github_pat_index+1,", total", num_token, "tokens are available.")
            #if false, retry until true, max number of retry is set to 3 times
            g = Github(access_token, retry = 3, timeout = 15)
            repo = g.get_repo(slug)
            stars = repo.stargazers_count
            watchers = repo.subscribers_count
            forks = repo.forks_count
            topics = repo.get_topics()
        except RateLimitExceededException as e:
          #  print(e.status)
          #  print('WARNING: Rate limit exceeded --', slug, ", using access token #", github_pat_index)
            #time.sleep(300)
            github_pat_index+=1
          #  print("***Exit current access token, proceed with next aceess token #", github_pat_index, "rescrape --",slug)
            break
        except BadCredentialsException as e:
           # print(e.status)
           # print('WARNING: Bad credentials exception --', slug, ", using access token #", github_pat_index)
            github_pat_index+=1
           # print("***Exit current access token, proceed with next aceess token #", github_pat_index, "rescrape --",slug)
            break
        except UnknownObjectException as e:
           # print(e.status)
           # print('WARNING: Unknown object exception --', slug)
            stars = None
            watchers = None
            forks =  None
            topics = None
            return slug, stars, watchers, forks, topics
            break
        except GithubException as e:
           # print(e.status)
           # print('General exception --', slug)
            break
        except r.exceptions.ConnectionError as e:
           # print('Retries limit exceeded --', slug)
           # print(str(e))
            time.sleep(10)
            continue
        except r.exceptions.Timeout as e:
           # print('Time out exception --', slug)
           # print(str(e))
            time.sleep(10)
            continue
            
        results = (slug, stars, watchers, forks, topics)
        #return slug, stars, watchers, forks, topics
        return results

# Example
i = 0
print(slugs_example[i])
result = pull_repo_stats(slugs_example[i], 0)
print(result[3])

#print(result[0])
#print(result[1] ==None)

## Multiprocessing approach: pool.apply_async for multiple parameters

def get_result(results):
    global slug_log
    global stars_log
    global watchers_log
    global forks_log
    global topics_log
    if result is None:
        slug_log.append(None)
        stars_log.append(None)
        watchers_log.append(None)
        forks_log.append(None)
        topics_log.append(None)  
    else:
        slug_log.append(results[0])
        stars_log.append(results[1])
        watchers_log.append(results[2])
        forks_log.append(results[3])
        topics_log.append(results[4])
        #print("appending results")

def get_stars(stars):
    global stars_log
    stars_log.append(stars)
    print("appending results")

stars_log = []
#practice on 2 slugs, one invalid slug 
#slugs_example = ["moderndive/ModernDive_book", "unknownrepo"]

#practice on 7 slugs, one invalid slug 
slugs_example = ["moderndive/ModernDive_book", "DSPG-Young-Scholars-Program/dspg21oss", 
                 "unknownrepo", "moderndive/ModernDive_book", "lsst/ip_diffim", "esrlabs/chipmunk", "paulmillr/chokidar"]


slugs_100 = slugs[0:100]

multiprocessing.cpu_count()

# Approach 2. pool.apply_async for multiple parameters
if __name__ == '__main__':
    cores_available = multiprocessing.cpu_count() - 1
    print(f'There are {cores_available} CPUs available.')
    #pool = multiprocessing.Pool(cores_available)
    pool = multiprocessing.Pool(10)
    
    # now we will feed in all of the remaining slugs 
    slug_log = []
    stars_log = []
    watchers_log = []
    forks_log = []
    topics_log = []

    start_time = datetime.datetime.now()
    print("Start scraping. Start time:", start_time)
    
    for slug in slugs_try:        
        pat_i = 3
        pool.apply_async(pull_repo_stats, args=(slug,  pat_i),callback = get_result)        
        print("~~~~~~~~~~~~~~~slug: ",slug, "using access token #", pat_i, "~~~~~~~~~~~~~~~~")
    pool.close()
    pool.join()
    
    final_log = pd.DataFrame({'slug': slug_log, "stars": stars_log, 'watchers': watchers_log, 'forks': forks_log, 'topics': topics_log}, columns=["slug", "stars", "watchers", "forks", "topics"])

    end_time = datetime.datetime.now()
    print("Finished scraping", len(final_log), "of", len(slugs_try), "records at", end_time)
    print("It took", end_time-start_time, "to run.")

#print(final_log)
final_log

## Multiprocessing approach: pool.imap_unordered

cores_available = multiprocessing.cpu_count() - 1
print(f'There are {cores_available} cores available.')
pool = Pool(cores_available)

# now we will feed in all of the remaining slugs 
slug_log = []
stars_log = []
watchers_log = []
forks_log = []
topics_log = []

start_time = datetime.datetime.now()
print("Start scraping. Start time:", start_time)
for result in pool.imap_unordered(pull_repo_stats, slugs_example):
    print("~~~~~~~~~~~~~~~result: ",result, "~~~~~~~~~~~~~~~~")
    if result is None:
        slug_log.append(None)
        stars_log.append(None)
        watchers_log.append(None)
        forks_log.append(None)
        topics_log.append(None)  
    else:
        slug_log.append(result[0])
        stars_log.append(result[1])
        watchers_log.append(result[2])
        forks_log.append(result[3])
        topics_log.append(result[4])


final_log = pd.DataFrame({'slug': slug_log, "stars": stars_log, 'watchers': watchers_log, 'forks': forks_log, 'topics': topics_log}, columns=["slug", "stars", "watchers", "forks", "topics"])

end_time = datetime.datetime.now()
print("Finished scraping", len(final_log), "of", len(slugs), "records at", end_time)
print("It took", end_time-start_time, "to run.")


final_log

#final_log.to_csv(r'/home/zz3hs/git/dspg21oss/data/dspg21oss/final_log_try.csv', index = False)   

# read in the file and check
print(final_log.head())
print(final_log)
print(final_log.isna().sum())
print(final_log.shape)

# concactinate scrapped dataframes together
df1 = pd.read_csv('/home/zz3hs/git/dspg21oss/data/dspg21oss/repo_stats_0707.csv') #import csv
print(df1.shape)

df2 = pd.read_csv('/home/zz3hs/git/dspg21oss/data/dspg21oss/new_repo_stats_8.csv') #import csv
df3 = pd.read_csv('/home/zz3hs/git/dspg21oss/data/dspg21oss/new_repo_stats_9.csv') #import csv

frames = [df1, df2, df3]

repo_stats_0707_new = pd.concat(frames)

repo_stats_0707_new = repo_stats_0707_new.drop_duplicates(subset = 'slug', keep = 'first')
print("We scrapped",repo_stats_0707_new.shape[0], "repos.")

#repo_stats_0707_new.to_csv(r'/home/zz3hs/git/dspg21oss/data/dspg21oss/repo_stats_0707_new.csv', index = False)
